+++ oap-server/server-storage-plugin/storage-banyandb-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/banyandb/BanyanDBStorageProvider.java
@@ -209,7 +209,10 @@ public class BanyanDBStorageProvider extends ModuleProvider {
                                                               BanyandbCommon.Metadata.newBuilder()
                                                                                      .setName(
                                                                                          BanyanDBUITemplateManagementDAO.GROUP))
-                                                          .setCatalog(BanyandbCommon.Catalog.CATALOG_UNSPECIFIED)
+                                                          .setCatalog(BanyandbCommon.Catalog.CATALOG_PROPERTY)
+                                                          .setResourceOpts(BanyandbCommon.ResourceOpts.newBuilder()
+                                                                                       .setShardNum(1)
+                                                                                       .build())
                                                           .build());
             this.modelInstaller.start();